resume-generator-portfolio/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
anthropic==0.25.0
weasyprint>=60.0
Markdown==3.5.1
